Add Record wrapper class
Update spec:setup to retrieve data for records, only if the given form
has at least one field (otherwise, MyTaskHelper will not allow records
to be created). Update FakeMTH and FixtureParser to parse record data.
Add records method to Form, which returns a lazy enumerator of the
associated fields. Since the results for records are paginated in the
MyTaskHelper API, a lazy enumerator is used to avoid unnecessary calls to
the API to get all pages.